+++ R/data.R
@@ -1,0 +1,12 @@
+#' Google Analytics API metadata
+#' 
+#' This is a local copy of the data provided by \link{google_analytics_meta}
+#' 
+#' Running your own call will be more up to date, but this is here in case.
+#' 
+#' It does not include the multi-channel or cohort variables.
+#' 
+#' @format A data frame with 476 rows and 15 variables
+#' 
+#' @source \url{https://developers.google.com/analytics/devguides/reporting/core/dimsmets}
+"meta"