+++ NEWS.md
@@ -1,5 +1,6 @@
-# 0.1.1.9000
+# 0.1.0.9000
 
+* segment helper as RStudio Addin or `gadget_GASegment()`
 * Fix order_type() for v4 (#18)
 * Tests
 * Multi-account batching for v3 (#17)
@@ -8,6 +9,7 @@
 # 0.1.1
 
 * Patch to fix ignoring first row of v4 API results (#16)
+>>>>>>> master
 
 # 0.1.0